Add attribute "marketplace_title" to VMIRelease
This commit introduces a new optional parameter on `VMIPushItem` model
called `marketplace_title_template` which is going to be used in
conjunction with `VMIPushItem.marketplace_title` property
by the pubtool as the marketplace "version title" whenever it's set, by
overriding the usual version title which the pubtool was going to
compute.

This is required for some layered products which requires a
specialized templating title instead of the one generated by the
pubtool.

With this, it will be possible to transport the template via the
VMIPushItem from the source to the pubtooling which will format it into
the apropriate value for the marketaplace.

Refers to SPSTRAT-116