ci: drop codecov upload in this project
In this project, "tox -e cov" already runs pytest with
`--cov-fail-under=100`, thus the coverage is always enforced to remain
at 100%.

Sending the coverage data to a remote service seems of little value
when the coverage is already enforced at 100% when the tests are run;
in fact, we probably wouldn't even reach this step if the coverage was
less than 100%, since tox should have failed.

It's a common source of CI flakiness, so just drop it.