Await dir availability while iterating StagedSource [RHELDST-25394]
scandir in `_push_items_for_topdir` was reporting path strings as
invalid arguments, presumably because they weren't yet available on
the OCP environment. This commit places a `wait_exist` call on the
path before doing anything with it. At the very least, we will now be
warned if a path is inaccessible prior to failure.