Let BootMode enum have a round-trippable repr
enum.Enum surprisingly violates the usual convention of having "repr" be
the inverse of "eval" (see bug [1]). For example, it produces a string
like <BootMode.hybrid: 'hybrid'>.

This breaks pushsource-ls command with the default black formatter, since
black expects the input to be valid Python code. Also, even if the
formatting worked, it would break the ability to copy-paste the
generated push items into a python script.

Fix it by overriding `__repr__`. The implementation is identical to the
suggestion in the stdlib docs[2].

[1] https://bugs.python.org/issue40066
[2] https://docs.python.org/3/library/enum.html#enum.Enum.__repr__