Fix pushsource-ls unable to serialize BootMode as YAML
pushsource-ls with YAML format enabled would crash if encountering a
BootMode field (as used on AMI push items). This is because pyyaml does
not enable any representer for enums by default.

Fix it by adding a custom representer which will simply serialize the
underlying value (e.g. BootMode.hybrid becomes "hybrid").

Note this is only being fixed for pushsource-ls and not in pushsource
library proper because there does not seem to be a single
universally-correct method of serializing enums into YAML. The chosen
representation here is merely something "good enough" for pushsource-ls
use-cases.