tests: improve pushsource-ls coverage
Test the pushsource-ls command over each staging area in the test data.
This helps ensure that any new fields on pushitems can be serialized
successfully. For example, it would have caught the serialization issue
with BootMode enum fixed in prior commits.

Requires adding 'black' to the test requirements so that this format can
be tested.