Add content reader to decouple NFS file reads [RHELDST-26339]
The API to read content of the push items strongly depends
on their presence as file on NFS or locally-mounted filesystem.
However, with advent of new push item sources, it likely that
the push items might not be stored as a file or be locally available.

Hence, add a new content reader interface that will be implemented
by content readers based on the source or content type. The readers
can then be used by specific push item types to get their bits from
the respective source.

Here, the File content reader implements the interface and is used
by default for all push items. This reader reads the bits from the
path provided as push item source and returns a file-like object as
push item content attribute.