Add content reader to decouple NFS file reads [RHELDST-26339]
Reading content of the push items strongly depends on their presence
as file on NFS or locally-mounted filesystem. However, with advent
of new push item sources, it likely that the push items might not be
stored as a file or be locally available.

Hence, this PR adds an `opener` to the pushitems that should fetch
the corresponding bit stream which will be accessible from `content()`.
`opener` should be defined in the `Source` while creating the pushitems
to get the bits from the specific content source. PushItem is not bound
to have an `opener` and `content()` will return the bits as file object
only when the `opener` is defined.

This adds a provision to transfer the responsibility of fetching the
content from the consumer that is using the content/pushitems to the
`Source` and corresponding `pushitem` here, abstracting the reading
mechanism from the user and provide flexibility to fetch from different
locations/protocols.